Kane object has new interfaces and documentation
The Kane object now takes constraints in a more consistent manner.
Kinematic differential equations and constraints are all provided as
equations equal to zero.
All the information in describing the system is now in a common form;
matrices and vectors. This will be written up in the subsubmodule documentation
using LaTeX in the future.
Some interfaces needed to be changed to accommodate this, and the corresponding
tests have been changed.